feat(chore): refactor left menu coin list
