--- packages/blockchain-wallet-v4-frontend/src/data/components/layoutWallet/selectors.js
@@ -1,7 +1,0 @@
-import { path } from 'ramda'
-
-export const getMenuOpened = path(['components', 'layoutWallet', 'menuOpened'])
-
-export const getSettingsOpened = path(['components', 'layoutWallet', 'settingsOpened'])
-
-export const getBalancesTable = path(['components', 'layoutWallet', 'balancesTable'])
