--- packages/blockchain-wallet-v4-frontend/src/data/components/layoutWallet/slice.js
@@ -1,22 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit'
-
-const initialState = {
-  menuOpened: false
-}
-
-const layoutWalletSlice = createSlice({
-  initialState,
-  name: 'layoutWallet',
-  reducers: {
-    layoutWalletMenuToggleClicked: (state) => {
-      state.menuOpened = !state.menuOpened
-    }
-  }
-})
-
-export const { layoutWalletMenuToggleClicked } = layoutWalletSlice.actions
-
-const { actions } = layoutWalletSlice
-const layoutWalletReducer = layoutWalletSlice.reducer
-
-export { actions, layoutWalletReducer }
